requirements.txt
torch
transformers
datasets
chromadb
beautifulsoup4
rank-bm25
nltk
openai
accelerate
sentence_transformers
peft
trl
liger-kernel
deepspeed
bitsandbytes
xformers
triton
torchvision
autoawq